requirements.txt
flask
python-dotenv
elevenlabs